Build with tsup and convert to ESM
